#6 - Added an errorHandler class that handles errors instead of having it in schemaValidator
